migration: add load classes
