+++ migration/load/postgresql/__init__.py
@@ -1,0 +1,5 @@
+from .postgresql import PostgreSQLCopyLoad
+
+__all__(
+    "PostgreSQLCopyLoad",
+)
