+++ migration/load/postgresql/tables/__init__.py
@@ -1,0 +1,7 @@
+from .records import RDMRecordTableLoad
+from .versions import RDMVersionStateComputedTable
+
+__all__(
+    "RDMRecordTableLoad",
+    "RDMVersionStateComputedTable",
+)
